+++ packages/ncloud-sdk/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@cloud-canvas/cloud-sdk",
+    "name": "@cloud-canvas/ncloud-sdk",
     "version": "0.0.0",
     "description": "",
     "private": false,
