+++ packages/ncloud-sdk/src/services/server/signature.ts
