+++ packages/ncloud-sdk/src/services/vpc/signature.ts
