+++ packages/ncloud-sdk/src/services/vserver/ApiClient.ts
@@ -1,4 +1,4 @@
-import { generateSignature } from "./signature";
+import { generateSignature } from './signature';
 interface ApiKey {
     accessKey?: string;
     secretKey?: string;
@@ -10,29 +10,35 @@ export class ApiClient {
     public basePath: string;
     public timeout: number;
 
-    constructor(apiKey: ApiKey){
+    constructor(apiKey: ApiKey) {
         this.apiKey = {};
-        if(apiKey.accessKey && apiKey.secretKey) this.apiKey = apiKey;
+        if (apiKey.accessKey && apiKey.secretKey) this.apiKey = apiKey;
         this.apiGWEndPoint = 'https://ncloud.apigw.ntruss.com';
-        if(process.env.NCLOUD_API_GW)
+        if (process.env.NCLOUD_API_GW)
             this.apiGWEndPoint = process.env.NCLOUD_API_GW;
-        this.basePath = (this.apiGWEndPoint + '/vserver/v2').replace(/\/+$/, '');
+        this.basePath = (this.apiGWEndPoint + '/vserver/v2').replace(
+            /\/+$/,
+            '',
+        );
         this.timeout = 60000;
     }
 
-    applyAuthToRequest(request: any, apiKey: ApiKey){
-        if(!apiKey.accessKey || !apiKey.secretKey)
+    applyAuthToRequest(request: any, apiKey: ApiKey) {
+        if (!apiKey.accessKey || !apiKey.secretKey)
             throw new Error(`Not Found AccessKey or SecretKey!`);
         const timestamp = Date.now();
         const ncloudHeader = {
             'x-ncp-apigw-timestamp': timestamp,
             'x-ncp-iam-access-key': apiKey.accessKey,
-            'x-ncp-apigw-signature-v1': generateSignature(request, timestamp, apiKey.accessKey, apiKey.secretKey)
-        }
+            'x-ncp-apigw-signature-v1': generateSignature(
+                request,
+                timestamp,
+                apiKey.accessKey,
+                apiKey.secretKey,
+            ),
+        };
         return ncloudHeader;
     }
 
-    async callAPI(){
-
-    }
+    async callAPI() {}
 }
