+++ packages/ncloud-sdk/src/signature.ts
